+++ src/Terrabuild.Configuration.Tests/Project.fs
@@ -73,6 +73,7 @@ let parseProject() =
               TargetBlock.Cache = None
               TargetBlock.Batch = Expr.Enum "partition" |> Some
               TargetBlock.Phase = None
+              TargetBlock.EnvironmentSensitive = Expr.Bool true |> Some
               TargetBlock.Steps = [ { Extension = "@dotnet"; Command = "build"; Parameters = Map.empty } ] }
         let targetDist =
             { TargetBlock.DependsOn = None
@@ -81,6 +82,7 @@ let parseProject() =
               TargetBlock.Cache = None
               TargetBlock.Batch = None
               TargetBlock.Phase = None
+              TargetBlock.EnvironmentSensitive = None
               TargetBlock.Steps = [ { Extension = "@dotnet"; Command = "build"; Parameters = Map.empty }
                                     { Extension = "@dotnet"; Command = "publish"; Parameters = Map.empty } ] }
         let targetDocker =
@@ -90,6 +92,7 @@ let parseProject() =
               TargetBlock.Cache = "remote" |> Expr.Enum |> Some
               TargetBlock.Batch = None
               TargetBlock.Phase = None
+              TargetBlock.EnvironmentSensitive = None
               TargetBlock.Steps = [ { Extension = "@shell"; Command = "echo"
                                       Parameters = Map [ "arguments", Expr.Function (Function.Trim,
                                                                                      [ Expr.Function (Function.Plus,
@@ -150,6 +153,7 @@ let parseProject2() =
               TargetBlock.Cache = None
               TargetBlock.Batch = None
               TargetBlock.Phase = None
+              TargetBlock.EnvironmentSensitive = None
               TargetBlock.Steps = [ { Extension = "@dotnet"; Command = "build"; Parameters = Map.empty } ] }
 
         let locals =