+++ src/Terrabuild.Tests/Core/Runner.fs
@@ -22,6 +22,7 @@ let private buildNode id projectDir target action operations =
       GraphDef.Node.ClusterHash = Some "cluster"
       GraphDef.Node.Operations = operations
       GraphDef.Node.EvaluationInputs = []
+      GraphDef.Node.EnvironmentSensitive = None
       GraphDef.Node.Artifacts = GraphDef.ArtifactMode.Workspace
       GraphDef.Node.Build = GraphDef.BuildMode.Auto
       GraphDef.Node.Batch = GraphDef.BatchMode.Single